Moving functions around, added new conditional for beginner preset, split weights file into weights and options
