+++ weights/beginner_override.json
@@ -1,28 +1,35 @@
 {
-    "shuffle_overworld_entrances": {
-        "false": 1
+    "options": {
+        "extra_conditionals": {
+            "restrict_one_entrance_randomizer": true
+        }
     },
-    "mq_dungeons": {
-        "0": 1
-    },
-    "hints": {
-        "always": 1
-    },
-    "hint_dist": {
-        "balanced": 1,
-        "strong": 1,
-        "tournament": 1,
-        "tournament_s3": 1,
-        "very_strong": 1
-    },
-    "damage_multiplier": {
-        "half": 1,
-        "normal": 2,
-        "double": 1
-    },
-    "item_pool_value": {
-        "plentiful": 1,
-        "balanced": 2,
-        "scarce": 1
+    "weights": {
+        "shuffle_overworld_entrances": {
+            "false": 1
+        },
+        "mq_dungeons": {
+            "0": 1
+        },
+        "hints": {
+            "always": 1
+        },
+        "hint_dist": {
+            "balanced": 1,
+            "strong": 1,
+            "tournament": 1,
+            "tournament_s3": 1,
+            "very_strong": 1
+        },
+        "damage_multiplier": {
+            "half": 1,
+            "normal": 2,
+            "double": 1
+        },
+        "item_pool_value": {
+            "plentiful": 1,
+            "balanced": 2,
+            "scarce": 1
+        }
     }
 }
