+++ weights/coop_override.json
@@ -1,14 +1,16 @@
 {
-    "bridge_tokens_max": 50,
-    "lacs_tokens_max": 50,
-    "triforce_goal_per_world_max": 50,
-    "mq_dungeons": {
-        "0": 1
-    },
-    "damage_multiplier": {
-        "normal": 1
-    },
-    "mq_dungeons_random": {
-        "False": 1
+    "weights": {
+        "bridge_tokens_max": 50,
+        "lacs_tokens_max": 50,
+        "triforce_goal_per_world_max": 50,
+        "mq_dungeons": {
+            "0": 1
+        },
+        "damage_multiplier": {
+            "normal": 1
+        },
+        "mq_dungeons_random": {
+            "False": 1
+        }
     }
 }
